requirements.txt
requests>=2.31.0
pandas>=2.0.0
pyinstaller>=5.13.2
pillow>=10.0.0
